+++ src/Ziggy.php
@@ -3,12 +3,14 @@
 namespace Tighten\Ziggy;
 
 use Illuminate\Contracts\Routing\UrlRoutable;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Reflector;
 use Illuminate\Support\Str;
 use JsonSerializable;
 use ReflectionClass;
 use ReflectionMethod;
+use ReflectionProperty;
 
 class Ziggy implements JsonSerializable
 {
@@ -67,14 +69,14 @@ class Ziggy implements JsonSerializable
             $filters = [];
 
             foreach ($group as $groupName) {
-                $filters = array_merge($filters, config("ziggy.groups.{$groupName}"));
+                $filters = array_merge($filters, Arr::wrap(config("ziggy.groups.{$groupName}")));
             }
 
-            return $this->filter($filters, true)->routes;
+            return $this->filter($filters)->routes;
         }
 
         if (config()->has("ziggy.groups.{$group}")) {
-            return $this->filter(config("ziggy.groups.{$group}"), true)->routes;
+            return $this->filter(config("ziggy.groups.{$group}"))->routes;
         }
 
         return $this->routes;
@@ -85,9 +87,33 @@ class Ziggy implements JsonSerializable
      */
     public function filter($filters = [], $include = true): self
     {
-        $this->routes = $this->routes->filter(function ($route, $name) use ($filters, $include) {
-            return Str::is(Arr::wrap($filters), $name) ? $include : ! $include;
+        $filters = Arr::wrap($filters);
+
+        $reject = collect($filters)->every(function (string $pattern) {
+            return Str::startsWith($pattern, '!');
         });
+
+        $this->routes = $reject
+            ? $this->routes->reject(function ($route, $name) use ($filters) {
+                foreach ($filters as $pattern) {
+                    if (Str::is(substr($pattern, 1), $name)) {
+                        return true;
+                    }
+                }
+            })
+            : $this->routes->filter(function ($route, $name) use ($filters, $include) {
+                if ($include === false) {
+                    return ! Str::is($filters, $name);
+                }
+
+                foreach ($filters as $pattern) {
+                    if (Str::startsWith($pattern, '!') && Str::is(substr($pattern, 1), $name)) {
+                        return false;
+                    }
+                }
+
+                return Str::is($filters, $name);
+            });
 
         return $this;
     }
@@ -107,19 +133,22 @@ class Ziggy implements JsonSerializable
 
         $bindings = $this->resolveBindings($routes->toArray());
 
-        return $routes->merge($fallbacks)
-            ->map(function ($route) use ($bindings) {
-                return collect($route)->only(['uri', 'methods', 'wheres'])
-                    ->put('domain', $route->domain())
-                    ->put('bindings', $bindings[$route->getName()] ?? [])
-                    ->when($middleware = config('ziggy.middleware'), function ($collection) use ($middleware, $route) {
-                        if (is_array($middleware)) {
-                            return $collection->put('middleware', collect($route->middleware())->intersect($middleware)->values()->all());
-                        }
+        $fallbacks->map(function ($route, $name) use ($routes) {
+            $routes->put($name, $route);
+        });
 
-                        return $collection->put('middleware', $route->middleware());
-                    })->filter();
-            });
+        return $routes->map(function ($route) use ($bindings) {
+            return collect($route)->only(['uri', 'methods', 'wheres'])
+                ->put('domain', $route->domain())
+                ->put('bindings', $bindings[$route->getName()] ?? [])
+                ->when($middleware = config('ziggy.middleware'), function ($collection) use ($middleware, $route) {
+                    if (is_array($middleware)) {
+                        return $collection->put('middleware', collect($route->middleware())->intersect($middleware)->values()->all());
+                    }
+
+                    return $collection->put('middleware', $route->middleware());
+                })->filter();
+        });
     }
 
     /**
@@ -173,8 +202,11 @@ class Ziggy implements JsonSerializable
                 $model = class_exists(Reflector::class)
                     ? Reflector::getParameterClassName($parameter)
                     : $parameter->getType()->getName();
-                $override = (new ReflectionClass($model))->isInstantiable()
-                    && $model === (new ReflectionMethod($model, 'getRouteKeyName'))->class;
+                $override = (new ReflectionClass($model))->isInstantiable() && (
+                    (new ReflectionMethod($model, 'getRouteKeyName'))->class !== Model::class
+                    || (new ReflectionMethod($model, 'getKeyName'))->class !== Model::class
+                    || (new ReflectionProperty($model, 'primaryKey'))->class !== Model::class
+                );
 
                 // Avoid booting this model if it doesn't override the default route key name
                 $bindings[$parameter->getName()] = $override ? app($model)->getRouteKeyName() : 'id';